Add welcome screen
This screen will show the application purpose and will be the first
screen user will see. It's on presentation module because wouldn't make
sense split it into a feature module as it's not an usable feature.

Face it like a Splash screen, it doesn't give user a feature, it just
shows us a little bit about our application brand.